Update of main.py
Changed the tag of rederror to 0xFF3030 and put the old one on the varibale rederror1. Also added the command /buy, the user need to put the item id to buy the item, you can edit it or add features to the shop so the user will be able to buy items more easyly with buttons.